refactor(gsd): extract discussion handoff boundaries
